+++ src/ResultScan.php
@@ -29,7 +29,6 @@ class ResultScan extends Collection
      * Ensures that each item is a `ScanEntity` and has all the properties it needs
      * @param \Cake\Collection\CollectionInterface|\Traversable|iterable $items Items
      * @return array<\LinkScanner\ScanEntity> Parsed items
-     * @throws \Tools\Exception\KeyNotExistsException
      */
     protected function parseItems(iterable $items): array
     {
@@ -45,7 +44,6 @@ class ResultScan extends Collection
     /**
      * Constructor
      * @param \Cake\Collection\CollectionInterface|\Traversable|iterable $items Items
-     * @throws \Tools\Exception\KeyNotExistsException
      */
     public function __construct($items = [])
     {
@@ -58,7 +56,6 @@ class ResultScan extends Collection
      * Returns a new `ResultScan` instance as the result of concatenating the list of elements.
      * @param iterable $items Items
      * @return \Cake\Collection\CollectionInterface
-     * @throws \Tools\Exception\KeyNotExistsException
      */
     public function append($items): CollectionInterface
     {
@@ -71,7 +68,6 @@ class ResultScan extends Collection
      * Returns a new `ResultScan` instance as the result of concatenating the passed list of elements.
      * @param mixed $items Items
      * @return \Cake\Collection\CollectionInterface
-     * @throws \Tools\Exception\KeyNotExistsException
      */
     public function prepend($items): CollectionInterface
     {