+++ src/Utility/LinkScanner.php
@@ -23,15 +23,13 @@ use Cake\Event\EventDispatcherTrait;
 use Cake\Event\EventList;
 use Cake\Http\Client;
 use Cake\Http\Client\Response;
-use ErrorException;
 use Exception;
 use Laminas\Diactoros\Stream;
 use LinkScanner\ResultScan;
 use LinkScanner\ScanEntity;
+use LogicException;
 use PHPUnit\Framework\Exception as PHPUnitException;
 use Serializable;
-use Tools\BodyParser;
-use Tools\Exceptionist;
 use Tools\Filesystem;
 
 /**
@@ -56,7 +54,7 @@ class LinkScanner implements Serializable
 
     /**
      * Default configuration
-     * @var array<string, mixed>
+     * @var array
      */
     protected array $_defaultConfig = [
         'cache' => true,
@@ -131,7 +129,7 @@ class LinkScanner implements Serializable
      */
     public function __get(string $name)
     {
-        return $this->$name;
+        return $this->{$name};
     }
 
     /**
@@ -162,14 +160,15 @@ class LinkScanner implements Serializable
     /**
      * Internal method to create a lock file
      * @return bool
-     * @throws \ErrorException
+     * @throws \LogicException
      */
     protected function _createLockFile(): bool
     {
-        Exceptionist::isTrue(
-            !$this->getConfig('lockFile') || !file_exists($this->lockFile),
-            __d('link-scanner', 'Lock file `{0}` already exists, maybe a scan is already in progress. If not, remove it manually', $this->lockFile)
-        );
+        if ($this->getConfig('lockFile') && file_exists($this->lockFile)) {
+            throw new LogicException(
+                __d('link-scanner', 'Lock file `{0}` already exists, maybe a scan is already in progress. If not, remove it manually', $this->lockFile)
+            );
+        }
 
         return $this->getConfig('lockFile') && Filesystem::instance()->createFile($this->lockFile);
     }
@@ -198,30 +197,30 @@ class LinkScanner implements Serializable
         $cacheKey = 'response_' . md5(serialize($url));
 
         //Tries to get the response from the cache
-        $response = $this->getConfig('cache') ? Cache::read($cacheKey, 'LinkScanner') : null;
-        if ($response && is_array($response)) {
-            [$response, $body] = $response;
+        $Response = $this->getConfig('cache') ? Cache::read($cacheKey, 'LinkScanner') : null;
+        if ($Response && is_array($Response)) {
+            [$Response, $body] = $Response;
 
-            $stream = new Stream('php://memory', 'wb+');
-            $stream->write($body);
-            $stream->rewind();
+            $Stream = new Stream('php://memory', 'wb+');
+            $Stream->write($body);
+            $Stream->rewind();
 
-            return $response->withBody($stream);
+            return $Response->withBody($Stream);
         }
 
         try {
-            $response = $this->Client->get($url);
+            $Response = $this->Client->get($url);
 
-            if ($this->getConfig('cache') && ($response->isOk() || $response->isRedirect())) {
-                Cache::write($cacheKey, [$response, (string)$response->getBody()], 'LinkScanner');
+            if ($this->getConfig('cache') && ($Response->isOk() || $Response->isRedirect())) {
+                Cache::write($cacheKey, [$Response, (string)$Response->getBody()], 'LinkScanner');
             }
         } catch (PHPUnitException $e) {
             throw $e;
         } catch (Exception $e) {
-            $response = (new Response())->withStatus(404);
+            $Response = (new Response())->withStatus(404);
         }
 
-        return $response;
+        return $Response;
     }
 
     /**
@@ -236,12 +235,11 @@ class LinkScanner implements Serializable
      * @param string $url Url to scan
      * @param string $referer Referer of this url
      * @return void
-     * @throws \Tools\Exception\KeyNotExistsException
      */
     protected function _recursiveScan(string $url, string $referer = ''): void
     {
-        $response = $this->_singleScan($url, $referer);
-        if (!$response) {
+        $Response = $this->_singleScan($url, $referer);
+        if (!$Response) {
             return;
         }
 
@@ -251,14 +249,14 @@ class LinkScanner implements Serializable
         }
 
         //Returns, if the response is not ok
-        if (!$response->isOk()) {
+        if (!$Response->isOk()) {
             $this->dispatchEvent('LinkScanner.responseNotOk', [$url]);
 
             return;
         }
 
         //Continues scanning for the links found
-        foreach ((new BodyParser($response->getBody(), $url))->extractLinks() as $link) {
+        foreach ((new BodyParser($Response->getBody(), $url))->extractLinks() as $link) {
             if (!$this->_canBeScanned($link)) {
                 continue;
             }
@@ -281,7 +279,6 @@ class LinkScanner implements Serializable
      * @param string $url Url to scan
      * @param string $referer Referer of this url
      * @return \Cake\Http\Client\Response|null
-     * @throws \Tools\Exception\KeyNotExistsException
      */
     protected function _singleScan(string $url, string $referer = ''): ?Response
     {
@@ -291,12 +288,12 @@ class LinkScanner implements Serializable
         }
 
         $this->dispatchEvent('LinkScanner.beforeScanUrl', [$url]);
-        $response = $this->_getResponse($url);
-        $location = $response->getHeaderLine('location');
-        $this->dispatchEvent('LinkScanner.afterScanUrl', [$response]);
+        $Response = $this->_getResponse($url);
+        $location = $Response->getHeaderLine('location');
+        $this->dispatchEvent('LinkScanner.afterScanUrl', [$Response]);
 
         //Follows redirects
-        if ($response->isRedirect() && $this->getConfig('followRedirects')) {
+        if ($Response->isRedirect() && $this->getConfig('followRedirects')) {
             if (!$this->_canBeScanned($location)) {
                 return null;
             }
@@ -309,12 +306,12 @@ class LinkScanner implements Serializable
         //Creates a `ScanEntity` instance with the result and appends it to the
         //  `ResultScan` instance
         $this->ResultScan = $this->ResultScan->appendItem(new ScanEntity([
-            'code' => $response->getStatusCode(),
+            'code' => $Response->getStatusCode(),
             'external' => is_external_url($url, $this->hostname),
-            'type' => $response->getHeaderLine('content-type'),
+            'type' => $Response->getHeaderLine('content-type'),
         ] + compact('location', 'url', 'referer')));
 
-        return $response;
+        return $Response;
     }
 
     /**
@@ -360,11 +357,13 @@ class LinkScanner implements Serializable
      * @param string|null $filename Filename where to export
      * @return string
      * @see serialize()
-     * @throws \ErrorException
+     * @throws \LogicException
      */
     public function export(?string $filename = null): string
     {
-        Exceptionist::isTrue(!$this->ResultScan->isEmpty(), __d('link-scanner', 'There is no result to export. Perhaps the scan was not performed?'));
+        if ($this->ResultScan->isEmpty()) {
+            throw new LogicException(__d('link-scanner', 'There is no result to export. Perhaps the scan was not performed?'));
+        }
 
         if ($this->getConfig('exportOnlyBadResults')) {
             $this->ResultScan = new ResultScan($this->ResultScan->filter(fn(ScanEntity $item): bool => $item->get('code') >= 400));
@@ -387,7 +386,7 @@ class LinkScanner implements Serializable
      *  - `LinkScanner.resultsImported`: will be triggered when the results have been exported.
      * @param string $filename Filename from which to import
      * @return $this
-     * @throws \ErrorException
+     * @throws \LogicException
      * @noinspection PhpMissingReturnTypeInspection
      */
     public function import(string $filename)
@@ -398,7 +397,7 @@ class LinkScanner implements Serializable
             $instance = unserialize(file_get_contents($filename) ?: '');
         } catch (Exception $e) {
             $message = preg_replace('/file_get_contents\([^)]+\):\s+/', '', $e->getMessage()) ?: '';
-            throw new ErrorException(
+            throw new LogicException(
                 __d('link-scanner', 'Failed to import results from file `{0}` with message `{1}`', $filename, lcfirst($message))
             );
         }
@@ -418,14 +417,16 @@ class LinkScanner implements Serializable
      *
      * Other events will be triggered by `_recursiveScan()` and `_singleScan()` methods.
      * @return $this
-     * @throws \ErrorException
+     * @throws \LogicException
      * @noinspection PhpMissingReturnTypeInspection
      */
     public function scan()
     {
         //Sets the full base url
         $fullBaseUrl = $this->getConfig('fullBaseUrl', Configure::read('App.fullBaseUrl') ?: 'http://localhost');
-        Exceptionist::isUrl($fullBaseUrl, __d('link-scanner', 'Invalid full base url `{0}`', $fullBaseUrl));
+        if (!is_url($fullBaseUrl)) {
+            throw new LogicException(__d('link-scanner', 'Invalid full base url `{0}`', $fullBaseUrl));
+        }
         $this->hostname = get_hostname_from_url($fullBaseUrl);
 
         $this->_createLockFile();